refactor(validation): 유효성 로직 분리
