+++ src/utils/validaition.js
@@ -1,62 +1,1 @@
-import { ERROR } from '../constants/errors.js';
-
-export const hasEmptyString = (input) => {
-  if (input === '') {
-    throw new Error(ERROR.IS_VALUE_EMPTY);
-  }
-};
-
-export const isValueInteger = (input) => {
-  if (!Number.isInteger(input)) {
-    throw new Error(ERROR.IS_NOT_POSITIVE_INTEGER);
-  }
-};
-
-export const validatePurchasePrice = (input) => {
-  const value = Number(input);
-
-  hasEmptyString(input);
-  isValueInteger(value);
-
-  if (value < 1000) {
-    throw new Error(ERROR.IS_LOWER_THAN_MINIMUM_OF_PUCHASE_PRICE);
-  }
-};
-
-const validateArrayOfWinningNumbers = (winningNumbers) => {
-  winningNumbers.forEach((value) => {
-    const winningNumber = Number(value);
-    hasEmptyString(value);
-    isValueInteger(winningNumber);
-
-    if (winningNumber < 1 || winningNumber > 45) {
-      throw new Error(ERROR.IS_NOT_RANGE_OF_WINNING_NUMBER);
-    }
-  });
-};
-
-export const validateWinningNumbers = (input) => {
-  const winningNumbers = input.split(',');
-  const winningNumberSet = new Set(winningNumbers);
-
-  if (winningNumberSet.size !== 6) {
-    throw new Error(ERROR.IS_DUPLICATED_WINNING_NUMBER);
-  }
-
-  if (winningNumbers.length !== 6) {
-    throw new Error(ERROR.IS_NOT_SAME_LENGTH_OF_WINNING_NUMBER);
-  }
-
-  validateArrayOfWinningNumbers(winningNumbers);
-};
-
-export const validateBonusNumber = (input, winningNumbers) => {
-  const bonusNumber = Number(input);
-
-  hasEmptyString(input);
-  isValueInteger(bonusNumber);
-
-  if (winningNumbers.includes(bonusNumber)) {
-    throw new Error(ERROR.IS_BONUS_NUMBER_DUPLICATED);
-  }
-};
+105,109,112,111,114,116,32,123,32,69,82,82,79,82,32,125,32,102,114,111,109,32,39,46,46,47,99,111,110,115,116,97,110,116,115,47,101,114,114,111,114,115,46,106,115,39,59,13,10,13,10,101,120,112,111,114,116,32,99,111,110,115,116,32,104,97,115,69,109,112,116,121,83,116,114,105,110,103,32,61,32,40,105,110,112,117,116,41,32,61,62,32,123,13,10,32,32,105,102,32,40,105,110,112,117,116,32,61,61,61,32,39,39,41,32,123,13,10,32,32,32,32,116,104,114,111,119,32,110,101,119,32,69,114,114,111,114,40,69,82,82,79,82,46,73,83,95,86,65,76,85,69,95,69,77,80,84,89,41,59,13,10,32,32,125,13,10,125,59,13,10,13,10,101,120,112,111,114,116,32,99,111,110,115,116,32,105,115,86,97,108,117,101,73,110,116,101,103,101,114,32,61,32,40,105,110,112,117,116,41,32,61,62,32,123,13,10,32,32,105,102,32,40,33,78,117,109,98,101,114,46,105,115,73,110,116,101,103,101,114,40,105,110,112,117,116,41,41,32,123,13,10,32,32,32,32,116,104,114,111,119,32,110,101,119,32,69,114,114,111,114,40,69,82,82,79,82,46,73,83,95,78,79,84,95,80,79,83,73,84,73,86,69,95,73,78,84,69,71,69,82,41,59,13,10,32,32,125,13,10,125,59,13,10,13,10
