+++ src/validation/validatePurchasePrice.js
@@ -1,0 +1,13 @@
+import { ERROR } from '../constants/errors.js';
+import { hasEmptyString, isValueInteger } from './validateInput.js';
+
+export const validatePurchasePrice = (input) => {
+  const value = Number(input);
+
+  hasEmptyString(input);
+  isValueInteger(value);
+
+  if (value < 1000) {
+    throw new Error(ERROR.IS_LOWER_THAN_MINIMUM_OF_PUCHASE_PRICE);
+  }
+};
